.github/scripts/requirements.txt
requests
azure-ai-ml
azure-identity
